Migrate to a fresh copy of the template

+++ app/localization/translateMissing.js
@@ -28,12 +28,12 @@ const {
 
   In order to use this file effectively, which is run with the command 'npm run translate', you would
   create translated strings like in menu.js or localization.jsx. You would then run the app and change
-  languages in order that the keys for these translated strings are populated in the various other
+  languages in order that the keys for these translated strings are populated in the various other 
   languages' missing.json files. Once this is done for all languages you'd like to create translations for, you may run `npm run translate` in order that the missing translation files be translated with
   the Google Translate API.
 
   Note - it is important that 'fromLanguage' be updated to the language that the keys are in the various
-  translation[.missing].json files. It is this variable that's used by Google to determine the source
+  translation[.missing].json files. It is this variable that's used by Google to determine the source 
   language from which to translate.
 */
 console.log("The translateMissing.js file must be updated before it can be ran.");
@@ -46,18 +46,12 @@ const translate = new Translate({
   projectId
 });
 
-function ourLocaleCodeToGoogleLocaleCode(locale) {
-  if (locale === 'fil') {
-    return 'tl';
-  }
-  return locale;
-}
-
 async function updateTranslations() {
+
   try {
     const root = "./app/localization/locales";
     const fromLanguage = "en";
-
+    
     // Get valid languages from Google Translate API
     let [googleLanguages] = await translate.getLanguages(); // ie. { code: "en", name: "English" }
     googleLanguages = googleLanguages.map(gl => gl.code.replace("-", "_"))
@@ -68,79 +62,70 @@ async function updateTranslations() {
     // Get all language directories;
     // https://stackoverflow.com/a/35759360/1837080
     const getDirectories = p => readdirSync(p).filter(f => statSync(join(p, f)).isDirectory());
-    const languageDirectories = getDirectories(root).filter(d => googleLanguages.includes(ourLocaleCodeToGoogleLocaleCode(d)));
-
-    const missingTranslationFile = `${root}/${fromLanguage}/translation.missing.json`;
-    const translationMissingExists = fs.existsSync(missingTranslationFile);
-
-    if (!translationMissingExists) {
-      console.error(`Could not generate translations because ${missingTranslationFile} does not exist, skipping!`);
-      return;
-    }
-
-    const missing = JSON.parse(fs.readFileSync(missingTranslationFile, {
-      encoding: "utf8"
-    }));
-
-    const missingKeys = Object.keys(missing);
-
-    if (!missingKeys.length) {
-      console.log(`Skipped creating translations; none found!`);
-      return;
-    }
-
+    const languageDirectories = getDirectories(root).filter(d => googleLanguages.includes(d));    
 
     // For each language, read in any missing translations
     // and translate
-    for (let i = 0; i < languageDirectories.length; i++) {
+    for (const languageDirectory of languageDirectories) {
+
       // Check to make sure each language has the proper files
       try {
-        const languageRoot = `${root}/${languageDirectories[i]}`;
+        const languageRoot = `${root}/${languageDirectory}`;
         const translationFile = `${languageRoot}/translation.json`;
-        const translationExists = fs.existsSync(translationFile);
-        const missingTranslationFileForCurrentLanguage = `${languageRoot}/translation.missing.json`;
+        const missingTranslationFile = `${languageRoot}/translation.missing.json`;
 
-        if (translationExists) {
+        const translationExists = fs.existsSync(translationFile);
+        const translationMissingExists = fs.existsSync(missingTranslationFile);
+
+        if (translationExists && translationMissingExists) {
+
           // Read in contents of files
-          let translations = JSON.parse(fs.readFileSync(translationFile, {
+          const translations = JSON.parse(fs.readFileSync(translationFile, {
+            encoding: "utf8"
+          }));
+          const missing = JSON.parse(fs.readFileSync(missingTranslationFile, {
             encoding: "utf8"
           }));
 
-          if (languageDirectories[i] === fromLanguage) {
-            for (let j = 0; j < missingKeys.length; j++) {
-              translations[missingKeys[j]] = missingKeys[j];
-            }
-          } else {
+          // Only translate files with actual values
+          const missingKeys = Object.keys(missing);        
+          if (missingKeys.length > 0){
+
             // Translate each of the missing keys to the target language
-            for (let j = 0; j < missingKeys.length; j++) {
-              const googleTranslation = await translate.translate(missingKeys[j], {
+            for (const missingKey of missingKeys){
+              const googleTranslation = await translate.translate(missingKey, {
                 from: fromLanguage,
-                to: ourLocaleCodeToGoogleLocaleCode(languageDirectories[i]),
+                to: languageDirectory
               });
 
               // Only set if a value is returned
-              if (googleTranslation.length > 0) {
-                translations[missingKeys[j]] = googleTranslation[0];
-              }
+              if (googleTranslation.length > 0){
+                translations[missingKey] = googleTranslation[0];
+              }              
             }
-          }
-
-          // Write output back to file
-          fs.writeFileSync(translationFile, JSON.stringify(translations, null, 2));
-          fs.writeFileSync(missingTranslationFileForCurrentLanguage, JSON.stringify({}, null, 2));
-
-          console.log(`Successfully updated translations for ${languageDirectories[i]}`);
+  
+            // Write output back to file
+            fs.writeFileSync(translationFile, JSON.stringify(translations, null, 2));
+            fs.writeFileSync(missingTranslationFile, JSON.stringify({}, null, 2));
+  
+            console.log(`Successfully updated translations for ${languageDirectory}`);
+          } else {
+            console.log(`Skipped creating translations for ${languageDirectory}; none found!`);
+          }          
         } else {
+
           // Log if we failed
-          console.error(`Could not generate translations for language '${languageDirectories[i]}' because ${translationFile} does not exist, skipping!`);
+          if (!translationExists) {
+            console.error(`Could not generate translations for language '${languageDirectory}' because ${translationFile} does not exist, skipping!`);
+          } else if (!translationMissingExists) {
+            console.error(`Could not generate translations for language '${languageDirectory}' because ${missingTranslationFile} does not exist, skipping!`);
+          }
         }
       } catch (error) {
         console.error("Failed due to fatal error");
         console.error(error);
       }
     }
-
-    fs.writeFileSync(missingTranslationFile, JSON.stringify({}, null, 2));
   } catch (e) {
     console.error("Failed due to fatal error");
     console.error(e);
